+++ packages/ak-avatar/src/index.js
@@ -1,23 +1,29 @@
-import headStyles from 'style!./host.less'; // eslint-disable-line no-unused-vars, import/no-unresolved, max-len
+/** @jsx vdom */
+import 'style!./host.less'; // eslint-disable-line import/no-unresolved
+
+import classNames from 'classnames';
 import shadowStyles from './shadow.less';
 
 import { enumeration } from 'akutil-common';
-import { define, vdom, state } from 'skatejs';
+import { vdom } from 'skatejs';
 
 const SIZE_ATTRIBUTE_ENUM = {
   attribute: 'size',
-  values: ['xsmall', 'small', 'medium', 'large', 'xlarge', 'xxlarge', 'xxxlarge'],
+  values: ['xsmall', 'small', 'medium', 'large', 'xxlarge'],
   missingDefault: 'medium',
   invalidDefault: 'medium',
 };
 
 const definition = {
   render(elem) {
-    vdom.style(shadowStyles.toString());
-    vdom.img({
-      alt: elem.alt,
-      src: elem.src || '',
-    });
+    const classes = classNames([shadowStyles.locals.img, shadowStyles.locals[elem.size]]);
+
+    return (
+      <span>
+        <style>{shadowStyles.toString()}</style>
+        <img alt={elem.alt} src={elem.src} class={classes} />
+      </span>
+    );
   },
   props: {
     size: enumeration(SIZE_ATTRIBUTE_ENUM)({
@@ -28,17 +34,9 @@ const definition = {
     },
     alt: {
       attribute: true,
+      default: '',
     },
-  },
-  ready(elem) {
-    state(elem, {
-      className: headStyles['ak-avatar'],
-    });
   },
 };
 
-
-/* The constructor for our component */
-export default () => define('ak-avatar', definition);
-
-export { definition };
+export default definition;