+++ packages/ak-avatar/stories/avatar-story.js
@@ -1,13 +1,13 @@
 import { storiesOf } from '@kadira/storybook';
 import reactify from 'akutil-react';
-import { definition } from '../src/index';
+import definition from '../src/index';
 import { define } from 'skatejs';
 import avatarStoryStyles from 'style!./stories.less'; // eslint-disable-line import/no-unresolved
 import { name } from '../package.json';
 
 const { React, ReactDOM, uniqueWebComponent } = window;
 
-const Avatar = reactify(uniqueWebComponent('ak-avatar', definition, define), {
+const Avatar = reactify(uniqueWebComponent(name, definition, define), {
   React,
   ReactDOM,
 });
@@ -15,6 +15,9 @@ const Avatar = reactify(uniqueWebComponent('ak-avatar', definition, define), {
 storiesOf(name, module)
   .add('A default avatar', () => (
     <Avatar src="//docs.atlassian.com/aui/latest/docs/images/avatar-96.png" />
+  ))
+  .add('An avatar with an incorrectly defined size (falls back to default)', () => (
+    <Avatar src="//docs.atlassian.com/aui/latest/docs/images/avatar-96.png" size="megalarge" />
   ))
   .add('A xsmall avatar', () => (
     <Avatar src="//docs.atlassian.com/aui/latest/docs/images/avatar-96.png" size="xsmall" />
@@ -31,7 +34,6 @@ storiesOf(name, module)
       <Avatar src={avatarUrl} size="small" style={avatarRowStyle} />
       <Avatar src={avatarUrl} size="medium" style={avatarRowStyle} />
       <Avatar src={avatarUrl} size="large" style={avatarRowStyle} />
-      <Avatar src={avatarUrl} size="xlarge" style={avatarRowStyle} />
       <Avatar src={avatarUrl} size="xxlarge" style={avatarRowStyle} />
     </div>);
   })