+++ packages/ak-inline-dialog/src/index.js
@@ -1,13 +1,14 @@
-import headStyles from 'style!./host.less'; // eslint-disable-line no-unused-vars, import/no-unresolved, max-len
-import shadowStyles from './shadow.less';
-import Layer from 'ak-layer'; // eslint-disable-line no-unused-vars
+/** @jsx vdom */
+/* eslint react/no-unknown-property: 0 */
+
 import { attachmentMap, getPositionFromClasses } from 'akutil-common';
-
-import { define, vdom, prop } from 'skatejs';
-
+import { vdom, prop } from 'skatejs';
+import headStyles from 'style!./host.less'; // eslint-disable-line import/no-unresolved, max-len
+import shadowStyles from './shadow.less';
 import webanimation from 'web-animations-js/web-animations-next.min'; // eslint-disable-line no-unused-vars, max-len
+import 'ak-layer';
 
-const Animations = {
+const Animations = { // eslint-disable-line no-unused-vars
   left: [
     { transform: 'translate3d(100%, 0, 0)', opacity: 1 },
     { transform: 'translate3d(0, 0, 0)', opacity: 1 },
@@ -46,64 +47,60 @@ const definition = {
     }
   },
   render(elem) {
-    if (elem.open) {
-      vdom.element('style', shadowStyles.toString());
-      vdom.element('ak-layer', {
-        position: elem.position,
-        target: elem.target,
-        movable: elem,
-        open: elem.open,
-        attachment: elem.attachment,
-      }, () => {
-        vdom.element('ak-animmytest', {
-          alignment: getAnimationPosition(elem),
-        }, () => {
-          const divAttrs = {
-            class: shadowStyles.locals.inlineDialogContainer,
-          };
-          vdom.element('div', divAttrs, () => {
-            vdom.element('slot');
-          });
-        });
-      });
-    }
+    return (
+      <ak-layer
+        open={elem.open}
+        position={elem.position}
+        attachment={elem.attachment}
+        target={elem.target}
+        renderElementTo={elem.renderElementTo}
+        renderElement={elem}
+      >
+        <ak-animtest alignment={getAnimationPosition(elem)} open={elem.open}>
+          <style>{shadowStyles.toString()}</style>
+          <div class={shadowStyles.locals.inlineDialogContainer}>
+            <slot />
+          </div>
+        </ak-animtest>
+      </ak-layer>
+    );
   },
   props: {
-    position: prop.string({ attribute: true }),
-    open: prop.boolean({ attribute: true }),
+    position: prop.string({ attribute: true, default: 'right middle' }),
+    open: prop.string({ attribute: true, default: 'false' }),
     target: prop.string({ attribute: true }),
     actualPosition: prop.string({ attribute: true }),
-    attachment: prop.string({ attribute: true }),
+    attachment: prop.string({ attribute: true, default: 'window' }),
+    renderElementTo: prop.string({ attribute: true }),
   },
 };
 
-define('ak-animmytest', {
+const AnimmyTestDefinition = { // eslint-disable-line no-unused-vars
   render(elem) {
-    let container;
-
-    vdom.element('div', {
+    const container = vdom.element('div', {
       class: shadowStyles.locals.animateContainer,
     }, () => {
-      container = vdom.element('div', {
-        class: shadowStyles.locals.animateContainer2,
-      }, () => {
+      vdom.element('div', {}, () => {
         vdom.element('slot');
       });
     });
     if (elem.alignment && Animations[elem.alignment]) {
-      container.animate(Animations[elem.alignment], {
+      container.firstChild.className = shadowStyles.locals.animateContainerInner;
+      container.firstChild.animate(Animations[elem.alignment], {
         duration: 200,
         iterations: 1,
       });
 
       setTimeout(() => {
-        container.className = '';
-      });
+        container.firstChild.className = '';
+      }, 200);
     }
   },
   props: {
     alignment: prop.string({ attribute: true }),
+    open: prop.string({ attribute: true }),
   },
-});
+};
 
-export default define('ak-inline-dialog', definition);
+export default definition;
+export { AnimmyTestDefinition };