+++ packages/ak-layer/src/index.register.js
@@ -1,0 +1,7 @@
+import { define } from 'skatejs';
+import definition from './index.js';
+
+// Expose the WebComponent
+define('ak-layer', definition);
+
+export default definition;