+++ packages/ak-avatar/CHANGELOG.md
@@ -1,5 +1,30 @@
-<a name="0.12.0"></a>
-# 0.12.0 (2016-07-04)
+<a name="0.12.7"></a>
+## [0.12.7](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/0.12.7-semver-tag-for-ak-avatar...v0.12.7) (2016-07-11)
+
+
+
+<a name="0.12.7-semver-tag-for-ak-avatar"></a>
+## [0.12.7-semver-tag-for-ak-avatar](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/0.12.6-semver-tag-for-ak-avatar...0.12.7-semver-tag-for-ak-avatar) (2016-07-11)
+
+
+### Bug Fixes
+
+* **sizes:** there are only 5 sizes ([ed9d5dc](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/ed9d5dc))
+
+
+
+<a name="0.12.6-semver-tag-for-ak-avatar"></a>
+## [0.12.6-semver-tag-for-ak-avatar](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/0.12.1-semver-tag-for-ak-avatar...0.12.6-semver-tag-for-ak-avatar) (2016-07-08)
+
+
+
+<a name="0.12.1-semver-tag-for-ak-avatar"></a>
+## [0.12.1-semver-tag-for-ak-avatar](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/b7f4571...0.12.1-semver-tag-for-ak-avatar) (2016-07-07)
+
+
+### Bug Fixes
+
+* Bump skatejs-named-slots to 0.2.8 ([b7f4571](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/b7f4571))