+++ packages/ak-inline-dialog/CHANGELOG.md
@@ -1,5 +1,19 @@
-<a name="45.0.0"></a>
-# 45.0.0 (2016-07-04)
+<a name="67.0.0"></a>
+# 67.0.0 (2016-07-11)
+
+
+### Bug Fixes
+
+* Bump skatejs-named-slots to 0.2.8 ([b7f4571](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/b7f4571))
+* fixed troubles with switching between stories ([d0fe678](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/d0fe678))
+* remove use of a uniqueWebComponentOld ([200899e](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/200899e))
+
+
+### Features
+
+* introducing register.js ([bc60ec1](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/bc60ec1))
+* new story ([30addc3](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/30addc3))
+* return using layer into inline-dialog ([9a905d2](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/9a905d2))