+++ packages/ak-layer/CHANGELOG.md
@@ -1,5 +1,16 @@
-<a name="43.0.0"></a>
-# 43.0.0 (2016-07-04)
+<a name="54.0.0"></a>
+# 54.0.0 (2016-07-11)
+
+
+### Bug Fixes
+
+* Bump skatejs-named-slots to 0.2.8 ([b7f4571](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/b7f4571))
+* fixed troubles with switching between stories ([d0fe678](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/d0fe678))
+
+
+### Features
+
+* return using layer into inline-dialog ([9a905d2](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/9a905d2))