+++ packages/akutil-common/CHANGELOG.md
@@ -1,5 +1,17 @@
-<a name="0.0.3"></a>
-## 0.0.3 (2016-07-04)
+<a name="0.1.0"></a>
+# 0.1.0 (2016-07-11)
+
+
+### Bug Fixes
+
+* Bump skatejs-named-slots to 0.2.8 ([b7f4571](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/b7f4571))
+* fixed troubles with switching between stories ([d0fe678](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/d0fe678))
+
+
+### Features
+
+* new story ([30addc3](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/30addc3))
+* return using layer into inline-dialog ([9a905d2](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/9a905d2))