+++ packages/akutil-component-template/CHANGELOG.md
@@ -1,5 +1,35 @@
-<a name="3.0.6"></a>
-## 3.0.6 (2016-07-04)
+<a name="3.1.1"></a>
+## [3.1.1](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/3.1.1-semver-tag-for-akutil-component-template...v3.1.1) (2016-07-11)
+
+
+
+<a name="3.1.1-semver-tag-for-akutil-component-template"></a>
+## [3.1.1-semver-tag-for-akutil-component-template](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/3.1.0-semver-tag-for-akutil-component-template...3.1.1-semver-tag-for-akutil-component-template) (2016-07-08)
+
+
+
+<a name="3.1.0-semver-tag-for-akutil-component-template"></a>
+# [3.1.0-semver-tag-for-akutil-component-template](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/3.0.7-semver-tag-for-akutil-component-template...3.1.0-semver-tag-for-akutil-component-template) (2016-07-08)
+
+
+### Bug Fixes
+
+* add URL to package.json ([e1b6b42](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/e1b6b42))
+
+
+### Features
+
+* add host.less example ([edabf91](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/edabf91))
+
+
+
+<a name="3.0.7-semver-tag-for-akutil-component-template"></a>
+## [3.0.7-semver-tag-for-akutil-component-template](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/b7f4571...3.0.7-semver-tag-for-akutil-component-template) (2016-07-07)
+
+
+### Bug Fixes
+
+* Bump skatejs-named-slots to 0.2.8 ([b7f4571](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/b7f4571))