+++ packages/akutil-polyfills/CHANGELOG.md
@@ -1,5 +1,17 @@
-<a name="0.0.4"></a>
-## 0.0.4 (2016-07-04)
+<a name="0.0.10"></a>
+## [0.0.10](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/0.0.10-semver-tag-for-akutil-polyfills...v0.0.10) (2016-07-11)
+
+
+
+<a name="0.0.10-semver-tag-for-akutil-polyfills"></a>
+## [0.0.10-semver-tag-for-akutil-polyfills](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/compare/d0fe678...0.0.10-semver-tag-for-akutil-polyfills) (2016-07-08)
+
+
+### Bug Fixes
+
+* bricked release 0.0.9 ([20604da](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/20604da))
+* Bump skatejs-named-slots to 0.2.8 ([b7f4571](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/b7f4571))
+* fixed troubles with switching between stories ([d0fe678](https://bitbucket.org/atlassian/https://bitbucket.org/atlassian/atlaskit/commits/d0fe678))